fix: handle dashboard eval breakdown sources
